Add ExecutionEngineId to skvbc SimpleRequest
This allows dispatching requests to different execution engines. Add it
to each generated request and print it for read and write requests
received in the TesterReplica.

Also add a new script for running TesterReplica and TesterClient for our
radio related tests. It generates new keys and concord configuration each
run because the old config used by the basic scripts is no longer valid.
This will keep it working with any changes to configuration.